pyproject.toml
[tool.poetry]
name = "idoven"
version = "1.0.0"
description = ""
authors = ["Gerard"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"

[tool.poetry.group.test.dependencies]
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
